refactor: enhance configuration management with new types and logging improvements
