+++ internal/config/log.go
@@ -1,0 +1,84 @@
+package config
+
+import (
+	"bytes"
+	"fmt"
+	"os"
+	"sync"
+
+	"github.com/fatih/color"
+	"github.com/pkg/errors"
+	log "github.com/sirupsen/logrus"
+)
+
+type cliFormatter struct {
+	noColor bool
+}
+
+func (f *cliFormatter) Format(entry *log.Entry) ([]byte, error) {
+	var buf bytes.Buffer
+	if entry.Level == log.InfoLevel {
+		fmt.Fprintf(&buf, "%s\n", entry.Message)
+		return buf.Bytes(), nil
+	}
+	var levelLabel string
+	if f.noColor {
+		levelLabel = fmt.Sprintf("%-5s", entry.Level.String())
+	} else {
+		switch entry.Level {
+		case log.WarnLevel:
+			levelLabel = color.YellowString("%-5s", entry.Level.String())
+		case log.ErrorLevel:
+			levelLabel = color.RedString("%-5s", entry.Level.String())
+		case log.DebugLevel:
+			levelLabel = color.CyanString("%-5s", entry.Level.String())
+		default:
+			levelLabel = fmt.Sprintf("%-5s", entry.Level.String())
+		}
+	}
+	fmt.Fprintf(&buf, "%s %s\n", levelLabel, entry.Message)
+	return buf.Bytes(), nil
+}
+
+var isStderrPiped = sync.OnceValue(func() bool {
+	fi, err := os.Stderr.Stat()
+	return err == nil && (fi.Mode()&os.ModeCharDevice) == 0
+})
+
+// IsStderrPiped reports whether os.Stderr is not connected to a terminal.
+// Result is cached after the first call.
+func IsStderrPiped() bool {
+	return isStderrPiped()
+}
+
+// ShouldJSONErrors returns true when errors must be emitted as structured JSON.
+func ShouldJSONErrors() bool {
+	return Cfg.OutputType == "json" || IsStderrPiped()
+}
+
+// SetUpLogs sets the log output and log level.
+func SetUpLogs() error {
+	log.SetFormatter(&cliFormatter{noColor: Cfg.NoColorOutput.Value})
+
+	// In JSON errors mode suppress logrus entirely — utils.WriteError is the sole stderr writer.
+	if Cfg.OutputType == "json" || IsStderrPiped() {
+		log.SetLevel(log.FatalLevel)
+		return nil
+	}
+
+	if Cfg.Quiet {
+		log.SetOutput(os.Stderr)
+		log.SetLevel(log.ErrorLevel)
+		return nil
+	}
+	if Cfg.Debug {
+		Cfg.Verbosity = "debug"
+	}
+	lvl, err := log.ParseLevel(Cfg.Verbosity)
+	if err != nil {
+		return errors.Wrap(err, "parsing log level")
+	}
+	log.SetOutput(os.Stderr)
+	log.SetLevel(lvl)
+	return nil
+}
